forward port 513843: Don't crash when non-existing incidences are canceled.
svn path=/trunk/KDE/kdepim/; revision=516078

+++ libkcal/scheduler.cpp
@@ -255,15 +255,17 @@ bool Scheduler::acceptCancel(IncidenceBase *incidence,ScheduleMessage::Status /*
 {
   bool ret = false;
   const IncidenceBase *toDelete = mCalendar->incidenceFromSchedulingID( incidence->uid() );
-  Event *even = mCalendar->event(toDelete->uid());
-  if (even) {
-    mCalendar->deleteEvent(even);
-    ret = true;
-  } else {
-    Todo *todo = mCalendar->todo(toDelete->uid());
-    if (todo) {
-      mCalendar->deleteTodo(todo);
+  if ( toDelete ) {
+    Event *even = mCalendar->event(toDelete->uid());
+    if (even) {
+      mCalendar->deleteEvent(even);
       ret = true;
+    } else {
+      Todo *todo = mCalendar->todo(toDelete->uid());
+      if (todo) {
+        mCalendar->deleteTodo(todo);
+        ret = true;
+      }
     }
   }
   deleteTransaction(incidence);
